+++ refinedstorage2-network/src/main/java/com/refinedmods/refinedstorage2/api/network/impl/node/detector/DetectorNetworkNode.java
@@ -1,6 +1,6 @@
 package com.refinedmods.refinedstorage2.api.network.impl.node.detector;
 
-import com.refinedmods.refinedstorage2.api.network.node.AbstractNetworkNode;
+import com.refinedmods.refinedstorage2.api.network.impl.storage.AbstractNetworkNode;
 import com.refinedmods.refinedstorage2.api.resource.ResourceKey;
 
 import javax.annotation.Nullable;