+++ refinedstorage2-network/src/main/java/com/refinedmods/refinedstorage2/api/network/impl/storage/StorageConfiguration.java
@@ -1,4 +1,4 @@
-package com.refinedmods.refinedstorage2.api.network.node;
+package com.refinedmods.refinedstorage2.api.network.impl.storage;
 
 import com.refinedmods.refinedstorage2.api.resource.ResourceKey;
 import com.refinedmods.refinedstorage2.api.resource.filter.FilterMode;
@@ -10,6 +10,18 @@ import org.apiguardian.api.API;
 @API(status = API.Status.STABLE, since = "2.0.0-milestone.2.4")
 public interface StorageConfiguration extends Priority {
     AccessMode getAccessMode();
+
+    default boolean isInsertOnly() {
+        return getAccessMode() == AccessMode.INSERT;
+    }
+
+    default boolean isExtractOnly() {
+        return getAccessMode() == AccessMode.EXTRACT;
+    }
+
+    boolean isVoidExcess();
+
+    void setVoidExcess(boolean voidExcess);
 
     void setAccessMode(AccessMode accessMode);