+++ refinedstorage2-platform-common/src/main/java/com/refinedmods/refinedstorage2/platform/common/grid/view/CompositeGridResourceFactory.java
@@ -4,7 +4,6 @@ import com.refinedmods.refinedstorage2.api.grid.view.GridResource;
 import com.refinedmods.refinedstorage2.api.grid.view.GridResourceFactory;
 import com.refinedmods.refinedstorage2.api.resource.ResourceAmount;
 import com.refinedmods.refinedstorage2.platform.api.support.registry.PlatformRegistry;
-import com.refinedmods.refinedstorage2.platform.api.support.resource.PlatformResourceKey;
 import com.refinedmods.refinedstorage2.platform.api.support.resource.ResourceType;
 
 import java.util.Optional;
@@ -18,12 +17,9 @@ public class CompositeGridResourceFactory implements GridResourceFactory {
 
     @Override
     public Optional<GridResource> apply(final ResourceAmount resourceAmount) {
-        if (!(resourceAmount.getResource() instanceof PlatformResourceKey platformResource)) {
-            return Optional.empty();
-        }
         return resourceTypeRegistry.getAll()
             .stream()
-            .flatMap(type -> type.toGridResource(platformResource, resourceAmount.getAmount()).stream())
+            .flatMap(type -> type.toGridResource(resourceAmount).stream())
             .findFirst();
     }
 }